Moved temp to cache_retry
