+++ calisphere/institution_views.py
@@ -3,14 +3,13 @@ from django.shortcuts import render
 from django.urls import reverse
 from django.http import Http404
 from . import constants
-from .cache_retry import json_loads_url
+from .cache_retry import json_loads_url, search_index
 from .search_form import CampusForm, RepositoryForm
 from .collection_views import Collection, get_rc_from_ids
 from .facet_filter_type import (
     CollectionFF, RepositoryFF)
 from django.apps import apps
 from django.conf import settings
-from .temp import search_index
 
 
 import math
